update
Learning & being able to continue growing through building new skills is my passion, this is what brings me to apply for Enterprise Technical support. A chance to leave a mark on this department, and the industry I choose to work in is my motive to find the best practices. Not to stop at just technical solutions, I embrace the continuous search for the best ways to communicate the resolution through the foundation of genuine concern for the right solution, and build on it credibility from transparency that doesn't sacrifice the brand.